+++ doc/CHANGELOG.md
@@ -8,6 +8,8 @@
 - `[Calendar]` Added calendar event add/update/remove via modal/API feature to calendar. ([#757](https://github.com/infor-design/enterprise-wc/pull/795))
 - `[Charts]` Added the ability to rotate x axis labels in axis charts like line and bar. ([#738](https://github.com/infor-design/enterprise-wc/pull/738))
 - `[Form]` Added new 'IdsForm' component. ([#357](https://github.com/infor-design/enterprise-wc/pull/357))
+- `[Hyperlink]` Added better support for `routerLink` for some frameworks. ([#861](https://github.com/infor-design/enterprise-wc/pull/861))
+
 
 ## 1.0.0-beta.1